requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
dashscope==1.23.2
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
dotenv==0.9.9
eval-type-backport==0.2.2
executing==2.2.0
faiss-cpu==1.11.0
fastapi==0.115.12
fastjsonschema==2.21.1
filelock==3.18.0
filetype==1.2.0
fonttools==4.58.0
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2025.5.1
ftfy==6.3.1
google-ai-generativelanguage
gradio==5.20.0
google-api-core==2.24.2
google-auth==2.40.1
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-generativeai==0.8.3
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.0
idna==3.10
ipykernel==6.29.5
ipython==9.2.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.6
jiter==0.9.0
joblib==1.5.1
json5==0.12.0
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.7.2
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter-server==2.15.0
jupyter-server-terminals==0.5.3
jupyterlab==4.4.2
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.15
kiwisolver==1.4.8
langchain-core==0.3.60
langchain-openai==0.3.17
langgraph==0.4.5
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.70
langsmith==0.3.42
markupsafe
matplotlib==3.10.3
matplotlib-inline==0.1.7
mistune==3.1.3
mpmath==1.3.0
multidict==6.4.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.4.2
notebook-shim==0.2.4
numpy==2.2.5
open-clip-torch==2.32.0
openai==1.78.0
opencv-python==4.11.0.86
orjson==3.10.18
ormsgpack==1.9.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
prometheus-client==0.21.1
prompt-toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic-core==2.33.2
pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
pytz==2025.2
pyyaml==6.0.2
pyzmq==26.4.0
qwen-agent==0.0.22
referencing==0.36.2
regex==2024.11.6
replicate==1.0.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.24.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
send2trash==1.8.3
setuptools==80.4.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
starlette==0.46.2
sympy==1.14.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.9.0
timm==1.0.15
tinycss2==1.4.0
torch==2.7.0
torchvision==0.22.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing-extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.4.0
uvicorn==0.34.2
python-multipart
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.14
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
langgraph-checkpoint-mongodb==0.1.0
motor==3.6.0
